Support terminating flag processing after # of args
Commands like "ssh" and have specific behaviors around additional
arguments, where exactly one argument is allowed and subsequent
arguments are not processed for flags.  For example:

    ssh user@host cat --some -I value

executes the command "cat" with "--some", "-I", and "value". pflags
currently attempts to handle '--some' and '-I' as arguments on a
command structured like this.

This commit adds a MaxArguments flag to FlagSet which terminates
flag parsing after N arguments are read (similar to '--'). It is
intended for use by commands that emulate 'ssh' or other fixed
argument length commands that expect to receive flags for subcommands
after their fixed length.