Better support golang flag short values
In golang flags -bob and --bob are the same thing. Not so in pflags. But
when a golang flag was just -b or --b we would still only support --b.
This was a little awkward to users.

If a golang flag was declared as a single character, support both the -b
and --b versions in the pflag version.