Do not eat ALL of the args if we hit a -test. arg
Right now if we hit a `-test.*` arg we swallow all remaining args. This causes
us to just eat that one arg and keep processing the rest. This does mean that
something like:
-test.coverprofile moo.cov myarg1
Will leave the moo.cov and myarg1 to be parsed by pflag. Previously neither
would be parsed by pflag. To get the 'expected' behavior use:
-test.coverprofile=moo.cov myarg1

resolves #597