Add a method to expose unknown flags back to the user
Currently it is possible to ignore unknown flags (with
`FlagSet.ParseErrorsWhitelist.UnknownFlags`) but there is no way to find out
what they were after the fact.

Add a method which registers a slice into which all unknown arguments will be
accumulated.

Note that unknown short arguments which appear in a group (e.g. `-uuu`) will be
exploded (e.g. into `-u -u -u`) in this slice. Also any following value is
associated with the final option, e.g. `["-uuu", "foo"]` is interpreted as
`["-u", "-u", "-u", "foo"]`. This is consistent with the statement:

    Single dashes signify a series of shorthand letters for flags. All but the
    last shorthand letter must be boolean flags.

This is the reason why the call to `stripUnknownFlagValue` in
`parseSingleShortArg` becomes conditional on this being the last short arg in
the batch.

Add code to the existing `TestIgnoreUnknownFlags` (which already has
comprehensive set of the varities of possible unknown flags) to check this.

Also fixed a small cut-and-paste-o in the failure message of the test (use of
`f.ParseAll`).

Signed-off-by: Ian Campbell <ijc@docker.com>