Prevent help output from escaping default values as go strings
When creating the help text for a flag, the default value shouldn't be
be excaped. It isn't clear to an end user that we would be escaping those values.

Instead we should be printing the actual value and letting the users decide
when/how to add escaping based on how they're utilzing it.

Fixes #346

Signed-off-by: John Schnake <jschnake@vmware.com>